+++ changelog.md
@@ -3,6 +3,7 @@
 ## master
 - Improve file structure
 - Move the Model Phpbb to PhpbbUser
+- Improve welcome email
 
 ## v0.1.0
 - First version